Move fixtures and reused variables to conftest
